Fix methodology of some calculations Changes to be committed: 	modified:   triangles_BG.cc
The calculation described below has optimized:
Let us consider we have N triangles and r is radius of circle, inscribed
into each of them.
Let r_average is average value of all radiuses.
If we have m triangles (from N) with radius wich is less or equal to
average value, i.e.

r <= r_average

In this case, number of triangles with radius, greater than r_average is

N-m

Thus can improve program, because it simplifying algorythm allot.
And in fact, this is the whole idea of the present commit.